refactor: create fragment for featured movies
