+++ app/src/main/java/com/will/moviedbapp/presentation/view/adapter/MovieAdapter.kt
@@ -27,8 +27,7 @@ class MovieAdapter(
         private val bind: MovieListTileBinding
     ) : RecyclerView.ViewHolder(bind.root) {
         fun bind(movie: Movie) {
-//            bind.moviePoster.load(AppConstants.Network.BASE_POSTER_URL + movie.posterPath)
-            bind.movieTitle.text = movie.title
+            bind.moviePoster.load(AppConstants.Network.BASE_POSTER_URL + movie.posterPath)
         }
     }
 }
