+++ app/src/main/java/com/will/moviedbapp/presentation/view/home/HomeActivity.kt
@@ -2,19 +2,11 @@ package com.will.moviedbapp.presentation.view.home
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.lifecycleScope
-import androidx.recyclerview.widget.LinearLayoutManager
-import com.will.moviedbapp.core.state.StateResult
+import com.will.moviedbapp.R
 import com.will.moviedbapp.databinding.ActivityHomeBinding
-import com.will.moviedbapp.presentation.model.HomeAction
-import com.will.moviedbapp.presentation.view.adapter.MovieAdapter
-import com.will.moviedbapp.presentation.viewmodel.HomeViewModel
-import kotlinx.coroutines.launch
-import org.koin.androidx.viewmodel.ext.android.viewModel
+import com.will.moviedbapp.presentation.view.home.fragments.FeaturedMoviesFragment
 
 class HomeActivity : AppCompatActivity() {
-
-    private val viewModel by viewModel<HomeViewModel>()
 
     private lateinit var binding: ActivityHomeBinding
 
@@ -23,35 +15,22 @@ class HomeActivity : AppCompatActivity() {
 
         initView()
         setListeners()
+        handleFragments()
 
         setContentView(binding.root)
     }
 
-    private fun setListeners() {
-        lifecycleScope.launch {
-            viewModel.trendingMovies.collect { state ->
-                when (state) {
-                    StateResult.Loading -> {}
-                    StateResult.Empty -> {}
-
-                    is StateResult.Success -> {
-                        binding.recyclerFeaturedMovies.setHasFixedSize(true)
-                        binding.recyclerFeaturedMovies.layoutManager =
-                            LinearLayoutManager(applicationContext)
-                        binding.recyclerFeaturedMovies.adapter = MovieAdapter(state.data)
-                    }
-
-                    is StateResult.Error -> {}
-                }
-            }
-        }
+    private fun handleFragments() {
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.trending_movies_frame_container, FeaturedMoviesFragment::class.java, null)
+            .commit()
     }
+
+    private fun setListeners() {}
 
     private fun initView() {
         binding = ActivityHomeBinding.inflate(layoutInflater)
         supportActionBar?.hide()
-
-        viewModel.handle(HomeAction.LoadTrendingMovies)
     }
 
 }
