+++ app/src/main/res/layout/activity_home.xml
@@ -56,26 +56,11 @@
 
     </com.google.android.material.textfield.TextInputLayout>
 
-    <TextView
-        android:id="@+id/text_featured_movies"
-        android:layout_width="0dp"
+    <FrameLayout
+        android:id="@+id/trending_movies_frame_container"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/spacing_xxl"
-        android:text="@string/featured_movies"
-        android:textColor="@color/white"
-        android:textFontWeight="700"
-        android:textSize="@dimen/font_body"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
+        android:layout_marginTop="@dimen/spacing_md"
         app:layout_constraintTop_toBottomOf="@+id/edt_container" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycler_featured_movies"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/spacing_lg"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/text_featured_movies" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
