+++ app/src/main/res/values/strings.xml
@@ -11,4 +11,6 @@
     <string name="greeting_subtitle">Book your favorite movies</string>
     <string name="search">Search</string>
     <string name="featured_movies">Featured Movies</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
